refactor: 잘못된 의존성 관계 변경
