+++ android/app/src/main/java/com/zzang/chongdae/domain/repository/OfferingRepository.kt
@@ -5,8 +5,8 @@ import com.zzang.chongdae.common.handler.Result
 import com.zzang.chongdae.domain.model.Filter
 import com.zzang.chongdae.domain.model.Meetings
 import com.zzang.chongdae.domain.model.Offering
+import com.zzang.chongdae.domain.model.OfferingWrite
 import com.zzang.chongdae.domain.model.ProductUrl
-import com.zzang.chongdae.presentation.view.write.OfferingWriteUiModel
 import okhttp3.MultipartBody
 
 interface OfferingRepository {
@@ -19,7 +19,7 @@ interface OfferingRepository {
         pageSize: Int?,
     ): Result<List<Offering>, DataError.Network>
 
-    suspend fun saveOffering(uiModel: OfferingWriteUiModel): Result<Unit, DataError.Network>
+    suspend fun saveOffering(offeringWrite: OfferingWrite): Result<Unit, DataError.Network>
 
     suspend fun saveProductImageOg(productUrl: String): Result<ProductUrl, DataError.Network>
